pyproject.toml
[project]
name = "cepa-backend"
version = "0.1.0"
description = "CEPA backend code"
authors = [
    {name = "jumashafara",email = "jumashafara0@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.6,<6.0.0",
    "djangorestframework>=3.16.1,<4.0.0",
    "django-cors-headers>=4.3.1,<5.0.0",
    "django-filter>=23.0.0,<24.0.0",
    "pillow>=10.0.0,<11.0.0"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
